Updated deltree task to not follow symlinks.
Submitted by: Jay Doane
Reviewed by: duncan


git-svn-id: https://svn.apache.org/repos/asf/tomcat@264230 13f79535-47bb-0310-9956-ffa450edef68

+++ archive/tools/trunk/ant/src/main/org/apache/tools/ant/taskdefs/Deltree.java
@@ -26,24 +26,34 @@ public class Deltree extends Task {
 		    " is not a dir";
 		throw new BuildException(msg);
 	    }
-	    removeDir(dir);
-	}
+            try {
+                removeDir(dir);
+            } catch (IOException ioe) {
+                String msg = "Unable to delete " + dir.getAbsolutePath();
+                throw new BuildException(msg);
+            }
+        }
     }
-
     
-    private void removeDir(File dir) {
-	String[] list = dir.list();
-	for (int i = 0; i < list.length; i++) {
-	    String s = list[i];
-	    File f = new File(dir, s);
-	    if (f.isDirectory()) {
-		removeDir(f);
-	    } else {
-		f.delete();
-	    }
-	}
-	dir.delete();
-    }
+    private void removeDir(File dir) throws IOException {
 
+        // check to make sure that the given dir isn't a symlink
+        // the comparison of absolute path and canonical path
+        // catches this
+        
+        if (dir.getCanonicalPath().equals(dir.getAbsolutePath())) {
+            String[] list = dir.list();
+            for (int i = 0; i < list.length; i++) {
+                String s = list[i];
+                File f = new File(dir, s);
+                if (f.isDirectory()) {
+                    removeDir(f);
+                } else {
+                    f.delete();
+                }
+            }
+        }
+        dir.delete();
+    }
 }
